Module Common

val core_types : Ppxlib.location -> (string * Ppxlib.expression) Base.List.t

val fold_right_type_params : (string -> 'a -> 'a) -> (Ppxlib.core_type * 'b) Base.List.t -> 'a -> 'a

val fold_right_type_decl : (string -> 'a -> 'a) -> Ppxlib.type_declaration -> 'a -> 'a

fold_right_type_decl fn accum type_ performs a right fold over all type variable (i.e. not wildcard) parameters in type_.

val poly_fun_of_type_decl : loc:Ppxlib__.Import.location -> Ppxlib.type_declaration -> Ppxlib__.Import.expression -> Ppxlib__.Import.expression

poly_fun_of_type_decl type_ expr wraps expr into fun poly_N -> ... for every type parameter 'N present in type_. For example, if type_ refers to type ('a, 'b) map, expr will be wrapped into fun poly_a poly_b -> [%e expr].

val fold_left_type_params : ('a -> string -> 'a) -> 'a -> (Ppxlib.core_type * 'b) Base.List.t -> 'a

val fold_left_type_decl : ('a -> string -> 'a) -> 'a -> Ppxlib.type_declaration -> 'a

fold_left_type_decl fn accum type_ performs a left fold over all type variable (i.e. not wildcard) parameters in type_.

val poly_apply_of_type_decl : loc:Ppxlib__.Location.t -> Ppxlib.type_declaration -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_407.Ast.Parsetree.expression -> Migrate_parsetree__Migrate_parsetree_versions.OCaml_407.Ast.Parsetree.expression

poly_apply_of_type_decl type_ expr wraps expr into expr poly_N for every type parameter 'N present in type_. For example, if type_ refers to type ('a, 'b) map, expr will be wrapped into [%e expr] poly_a poly_b. _ parameters are ignored.

val expr_of_option : loc:Ppxlib.location -> Ppxlib.expression option -> Ppxlib.expression

expr_of_option ~loc o turns an optional expression into an expression of an optional value. In several places there are optional attributes, e.g. @@version foo, which end up as values of type `expression option`. These are often turned into optional values in the generated code.

module Attrs : sig ... end

Typed attribute getters

val attr : Ppxlib__.Location.t -> Base.String.t -> (Base.String.t Ppxlib.loc * Ppxlib__.Import.payload) Base.List.t -> string Base.Option.t

val split : Base.String.t -> Base.String.t list

val convert_doc : Base.String.t -> Base.String.t Base.List.t

val get_doc : loc:Ppxlib__.Location.t -> Ppxlib__.Import.expression option -> Ppxlib.attributes -> Ppxlib__.Import.expression

get_doc loc rpcdoc attrs extracts documentation from the type declarations. rpcdoc is the result of looking for @@doc tags. If this is found, we use that. If not, we look for ocamldoc docstrings and return them instead. In both cases, the result is an expression of type list